Paper-select is now also wrapped in md-input-container as default which makes it compatible with paper-input. Uses paper-input has base.
