+++ addon/components/paper-select-value.js
@@ -5,11 +5,11 @@ export default Ember.Component.extend({
   classNames: ['md-select-value'],
   classNameBindings: ['isPlaceholder:md-select-placeholder'],
 
-  isPlaceholder: Ember.computed('value', function () {
+  isPlaceholder: Ember.computed('value', function() {
     return !this.get('value');
   }),
 
-  label: Ember.computed('isPlaceholder', function () {
+  label: Ember.computed('isPlaceholder', function() {
     if (this.get('isPlaceholder')) {
       return this.get('placeholder');
     } else {
