+++ app/components/paper-select-core.js
@@ -1,0 +1,2 @@
+import PaperSelectCore from 'ember-paper/components/paper-select-core';
+export default PaperSelectCore;
