+++ app/templates/components/paper-select-core.hbs
@@ -1,0 +1,14 @@
+{{paper-select-value placeholder=placeholder value=label}}
+{{#if isOpen}}
+  {{#paper-select-container}}
+    {{#paper-select-menu}}
+      {{#paper-content class="md-default-theme"}}
+        {{#if isLoading}}
+          <div>{{paper-progress-circular}}</div>
+        {{else}}
+          {{yield items}}
+        {{/if}}
+      {{/paper-content}}
+    {{/paper-select-menu}}
+  {{/paper-select-container}}
+{{/if}}
