+++ app/templates/components/paper-select.hbs
@@ -1,14 +1,11 @@
-{{paper-select-value placeholder=placeholder value=label}}
-{{#if isOpen}}
-  {{#paper-select-container}}
-    {{#paper-select-menu}}
-      {{#paper-content class="md-default-theme"}}
-        {{#if isLoading}}
-          <div>{{paper-progress-circular}}</div>
-        {{else}}
-          {{yield items}}
-        {{/if}}
-      {{/paper-content}}
-    {{/paper-select-menu}}
-  {{/paper-select-container}}
+{{#if label}}
+  <label class={{if model 'md-static' 'md-placeholder'}} for={{inputElementId}}>{{label}}</label>
 {{/if}}
+
+{{#if icon}}
+  {{paper-icon icon=icon class=icon-class}}
+{{/if}}
+
+{{#paper-select-core placeholder=placeholder model=model disabled=disabled on-open=onOpen item-label-callback=itemLabelCallback as |items|}}
+  {{yield items}}
+{{/paper-select-core}}
