+++ tests/dummy/app/templates/menu.hbs
@@ -7,7 +7,7 @@
   </div>
 {{/paper-toolbar}}
 
-{{#paper-content classNames="md-padding"}}
+{{#paper-content class="md-padding"}}
 <div class="doc-content">
